Eac dev (#26)
* protected against missing data

* Fixes to run ptc and nonlinearity with missing data

* Added teststand argument to eo_link_bot_files

* skip plotting if extract() failed, and warning message

* Added options for nonlinearity

* Added functions to define guard values in tables

* Added linking of detector response files

* Warn and skip instead of assert when exposure times don't match in flat pairs

* added parameters to spline creation, handle missing data in nonlinearity

* Added eo_link_run.py to link results from one run to another

* remove spurious print statements

* Fix error handling in nonlinearity

* delinting

* Added retry option to analysis tasks

* Added corner rafts and moved log location

* Protect outlier stat plotting against missing data

* added bright_pixel_mask and EOTEST results to set of files we make links for with bot runs

* Protect against missing data in SuperXTasks

* Fix bot_glob_imagetype to by flat0 and flat1 to match filenames

* Make batch log directory if it doesn't already exist

* Fix up BiasFFTTask, remove print statements

* Fix up photodiode data access in flat_pair.py

* Delinting

* delinting

* Fix call to histogram_array

* fixes for using footprints to detect defects

* remove spurious line from superdark.py

* Lots of changes to deal with nine raft data

* Lots of Updates

* changed to handle stability and new calibration management

* delinting

* Fix docstring

* Protect against missing dark_config.fits

* change text name to make it easier to select

* Fix test_file_utils_get_names_dc

* Protect against missing seq_list.fits

* Protect against missing seq_list.fits

* remove spurios print

* delinting

* Updated summary html to include dataset

* Various changes

* Added focal plane level amp results plotting

* delinting

* fix unit tests

* added templates/eo_html_report.yaml and minor changes to report_utils.py

+++ python/lsst/eo_utils/base/report_utils.py
@@ -4,6 +4,8 @@
 
 import sys
 
+import glob
+
 import os
 
 import shutil
@@ -16,8 +18,7 @@ import yaml
 
 from .defaults import EO_PACKAGE_BASE, ALL_SLOTS, NINE_RAFTS
 
-from .file_utils import makedir_safe,\
-    get_raft_names_dc, read_runlist
+from .file_utils import makedir_safe, read_runlist
 
 
 def get_report_config_info(table_tag, **kwargs):
@@ -58,11 +59,11 @@ def get_report_config_info(table_tag, **kwargs):
         table_desc = {}
 
     return dict(cssfile=cssfile,
-                table_desc=template_dict.get(table_tag, {}),
+                table_desc=table_desc,
                 defaults=template_dict['defaults'])
 
 
-def handle_file(file_name, outdir, action):
+def handle_file(file_name, outdir, action, overwrite=False):
     """Move, copy or link a file to an output directory
 
     Parameters
@@ -73,6 +74,8 @@ def handle_file(file_name, outdir, action):
         The output directory
     action : `str`
         What to do with the file, of `copy`, `move`, or `link`
+    overwrite : `bool`
+        Overwrite existing files
 
     Returns
     -------
@@ -85,10 +88,16 @@ def handle_file(file_name, outdir, action):
 
     inname = os.path.abspath(file_name)
     outname = os.path.abspath(os.path.join(outdir, basename))
-    try:
-        os.unlink(outname)
-    except FileNotFoundError:
-        pass
+
+    if overwrite:
+        try:
+            os.unlink(outname)
+        except FileNotFoundError:
+            pass
+    else:
+        if os.path.exists(outname):
+            print("File %s exists, skipping" % outname)
+            return basename
 
     makedir_safe(outname)
     if action in ['copy', 'cp']:
@@ -200,6 +209,8 @@ def create_plot_table_row(tbody_node, desc, plot_file, outdir, **kwargs):
     --------
     plot_report_action : `str`
         What to do with the plot file, of `copy`, `move`, or `link`
+    overwrite : `bool`
+        Overwrite existing files
     row_class : `str`
         The style class to use for the row node
     col_desc_class : `str`
@@ -214,12 +225,14 @@ def create_plot_table_row(tbody_node, desc, plot_file, outdir, **kwargs):
     row_node : `xml.etree.ElementTree.SubElement`
         The row node
     """
-
-    if not os.path.exists(plot_file):
+    plot_files = glob.glob(plot_file)
+    if not plot_files:
         sys.stdout.write("Warning, skipping missing plot %s\n" % plot_file)
         return None
 
-    basename = handle_file(plot_file, outdir, kwargs.get('plot_report_action', 'link'))
+    basename = handle_file(plot_files[0], outdir,
+                           kwargs.get('plot_report_action', 'link'),
+                           kwargs.get('overwrite', False))
 
     row_node = make_child_node(tbody_node, 'tr', node_class=kwargs.get('row_class', None))
     make_child_node(row_node, 'td',
@@ -261,6 +274,7 @@ def create_slot_table(parent_node, **kwargs):
         The table node
     """
     kwcopy = kwargs.copy()
+    prefix = kwcopy.get('prefix', '')
 
     html_file = kwargs.get('html_file', None)
     if html_file is not None:
@@ -282,7 +296,7 @@ def create_slot_table(parent_node, **kwargs):
     nslot = 0
     for slot in ALL_SLOTS:
         if basedir is not None:
-            slot_path = os.path.join(basedir, "%s.html" % slot)
+            slot_path = os.path.join(basedir, "%s%s.html" % (prefix, slot))
             if not os.path.exists(slot_path):
                 continue
 
@@ -295,9 +309,10 @@ def create_slot_table(parent_node, **kwargs):
                                    node_class=kwcopy.get('table_col_class', None))
         make_child_node(col_node, 'a',
                         text=slot,
-                        href="%s.html" % slot)
+                        href="%s%s.html" % (prefix, slot))
 
     if not nslot:
+        print("No slot data in %s, skipping" % basedir)
         parent_node.remove(h3_node)
         parent_node.remove(table_node)
         return None
@@ -330,15 +345,17 @@ def create_raft_table(parent_node, **kwargs):
         The table node
     """
     kwcopy = kwargs.copy()
+    prefix = kwcopy.get('prefix', '')
 
     html_file = kwargs.get('html_file', None)
     rafts = kwargs.get('rafts', NINE_RAFTS)
+    h3_text = kwargs.get('h3_text', 'List of rafts')
     if html_file is not None:
         basedir = os.path.dirname(html_file)
     else:
         basedir = None
 
-    h3_node = make_child_node(parent_node, 'h3', text="List of RAFTS")
+    h3_node = make_child_node(parent_node, 'h3', text=h3_text)
 
     table_node = make_child_node(parent_node, 'table')
     tbody_node = make_child_node(table_node, 'tbody')
@@ -352,7 +369,7 @@ def create_raft_table(parent_node, **kwargs):
     nraft = 0
     for raft in rafts:
         if basedir is not None:
-            raft_path = os.path.join(basedir, raft, "index.html")
+            raft_path = os.path.join(basedir, raft, "%sindex.html" % prefix)
             if not os.path.exists(raft_path):
                 continue
 
@@ -365,7 +382,7 @@ def create_raft_table(parent_node, **kwargs):
                                    node_class=kwcopy.get('table_col_class', None))
         make_child_node(col_node, 'a',
                         text=raft,
-                        href=os.path.join(raft, "index.html"))
+                        href=os.path.join(raft, "%sindex.html" % prefix))
 
     if not nraft:
         parent_node.remove(h3_node)
@@ -416,7 +433,7 @@ def create_plot_table(parent_node, table_desc, inputdir, outdir, **kwargs):
                     text="Plot",
                     node_class=header_col_class)
 
-    rowlist = table_desc['rows']
+    rowlist = table_desc.get('rows', [])
     nrows = 0
     for row_desc in rowlist:
         plotfile = os.path.join(inputdir, row_desc['figure'].format(**dataid))
@@ -482,6 +499,13 @@ def create_run_table(parent_node, dataset, **kwargs):
     """
     kwcopy = kwargs.copy()
 
+    html_file = kwcopy.get('html_file', None)
+    if html_file is not None:
+        basedir = os.path.dirname(html_file)
+    else:
+        basedir = None
+
+
     runlist = read_runlist("%s_runs.txt" % dataset)
 
     make_child_node(parent_node, 'h3', text="List of runs")
@@ -503,15 +527,36 @@ def create_run_table(parent_node, dataset, **kwargs):
                                    node_class=kwcopy.get('table_row_class', None))
         col_run_node = make_child_node(row_node, 'td',
                                        node_class=kwcopy.get('table_col_class', None))
-        raft = run_info[0].replace('-Dev', '')
-        run_url = os.path.join(run_info[1], raft, 'index.html')
 
-        make_child_node(col_run_node, 'a',
-                        href=run_url,
-                        text=run_info[1])
-        make_child_node(row_node, 'td',
-                        text=raft,
-                        node_class=kwcopy.get('table_col_class', None))
+        run_url = os.path.join(run_info[1], 'index.html')
+        if basedir is None or os.path.exists(os.path.join(basedir, run_url)):
+            make_child_node(col_run_node, 'a',
+                            href=run_url,
+                            text=run_info[1])
+
+
+        col_rafts_node = make_child_node(row_node, 'td',
+                                         node_class=kwcopy.get('table_col_class', None))
+
+        raft = run_info[0].replace('-Dev', '')
+
+        if raft in ['Cryostat-0001']:
+            rafts = NINE_RAFTS
+        else:
+            rafts = [raft]
+
+        nrafts = 0
+        for _raft in rafts:
+            raft_run_url = os.path.join(run_info[1], _raft, 'index.html')
+            if basedir is None or os.path.exists(os.path.join(basedir, raft_run_url)):
+                make_child_node(col_rafts_node, 'a',
+                                href=raft_run_url,
+                                text=_raft)
+                nrafts += 1
+
+        if not nrafts:
+            tbody_node.remove(row_node)
+
 
     return table_node
 
@@ -621,6 +666,7 @@ def write_raft_report(dataid, inputbase, outbase, **kwargs):
         outdir = os.path.join(outbase, dataid['run'], dataid['raft'])
         html_file = os.path.join(outdir, 'index.html')
 
+    kwcopy['html_file'] = html_file
     html_node = ET.Element('html')
     create_report_header(html_node,
                          title="Results for {run}:{raft}".format(**dataid),
@@ -630,9 +676,16 @@ def write_raft_report(dataid, inputbase, outbase, **kwargs):
 
     ntables = create_plot_tables(body_node, config_info['table_desc'], inputbase, outdir, **kwcopy)
 
-    kwcopy['html_file'] = html_file
-    if ntables:
-        create_slot_table(body_node, **kwcopy)
+    for slot in ALL_SLOTS:
+        dataid_slot = dataid.copy()
+        dataid_slot['slot'] = slot
+        kwcopy.pop('dataid', None)
+        write_slot_report(dataid_slot, inputbase, outbase, **kwcopy)
+
+    kwcopy['dataid'] = dataid
+    slot_table_node = create_slot_table(body_node, **kwcopy)
+
+    if ntables or slot_table_node is not None:
         makedir_safe(html_file)
         _ = handle_file(config_info['cssfile'], outdir, action='copy')
         write_tree_to_html(html_node, html_file)
@@ -641,13 +694,14 @@ def write_raft_report(dataid, inputbase, outbase, **kwargs):
 
 
 
-def write_run_report(run, inputbase, outbase, **kwargs):
+
+def write_run_report(data, inputbase, outbase, **kwargs):
     """Create table with descriptions and a plots
 
     Parameters
     ----------
-    run : `str`
-        Run number
+    dataid : `dict`
+        Dictionary with run, raft
     inputbase : `str`
         Input base directory
     outbase : `str` or `None`
@@ -656,47 +710,157 @@ def write_run_report(run, inputbase, outbase, **kwargs):
     Keywords
     --------
     """
-    sys.stdout.write("Writing report for %s\n" % run)
     kwcopy = kwargs.copy()
 
+
     config_info = get_report_config_info('run_plot_tables', **kwcopy)
+
     kwcopy.update(config_info['defaults'])
-    dataid = dict(run=run)
+    dataid = dict(run=data)
     kwcopy['dataid'] = dataid
 
     if outbase is None:
         outdir = None
         html_file = None
     else:
-        outdir = os.path.join(outbase, run)
+        outdir = os.path.join(outbase, dataid['run'])
         html_file = os.path.join(outdir, 'index.html')
-        makedir_safe(html_file)
-        ccsfile_out = handle_file(config_info['cssfile'], outdir, action='copy')
+
+    kwcopy['html_file'] = html_file
+    sys.stdout.write("Writing report for {run}\n".format(**dataid))
 
     html_node = ET.Element('html')
     create_report_header(html_node,
                          title="Results for {run}".format(**dataid),
+                         stylesheet=kwcopy.pop('stylesheet', os.path.basename(config_info['cssfile'])))
+
+    body_node = make_child_node(html_node, 'body')
+
+    ntables = create_plot_tables(body_node, config_info['table_desc'], inputbase, outdir, **kwcopy)
+
+    for raft in NINE_RAFTS:
+        dataid_raft = dataid.copy()
+        dataid_raft['raft'] = raft
+        kwcopy.pop('dataid', None)
+        write_raft_report(dataid_raft, inputbase, outbase, **kwcopy)
+
+    kwcopy['dataid'] = dataid
+    raft_table_node = create_raft_table(body_node, **kwcopy)
+
+    if ntables or raft_table_node is not None:
+        makedir_safe(html_file)
+        _ = handle_file(config_info['cssfile'], outdir, action='copy')
+        write_tree_to_html(html_node, html_file)
+    else:
+        sys.stdout.write("No data, skipping raft\n")
+
+
+
+
+def write_summary_report_by_slot(dataset, raft, slot, inputbase, outbase, **kwargs):
+    """Create table with descriptions and a plots
+
+    Parameters
+    ----------
+    dataset : `str`
+        Dataset id
+    raft : `str`
+        Raft id
+    slot : `str`
+        Slot id
+    inputbase : `str`
+        Input base directory
+    outbase : `str` or `None`
+        Output directory
+
+    """
+    kwcopy = kwargs.copy()
+
+    sys.stdout.write("Writing report for %s\n" % dataset)
+
+    config_info = get_report_config_info('summary_by_slot_plot_tables', **kwcopy)
+
+    kwcopy.update(config_info['defaults'])
+
+    if outbase is None:
+        outdir = None
+        html_file = None
+    else:
+        outdir = os.path.join(outbase, raft)
+        html_file = os.path.join(outdir, '%s_%s.html' % (dataset, slot))
+        makedir_safe(html_file)
+        ccsfile_out = handle_file(config_info['cssfile'], outdir, action='copy')
+
+    kwcopy['html_file'] = html_file
+    sys.stdout.write("Writing report for %s:%s:%s to %s\n" % (dataset, raft, slot, html_file))
+
+    html_node = ET.Element('html')
+    create_report_header(html_node,
+                         title="%s %s %s results" % (dataset, raft, slot),
                          stylesheet=kwcopy.pop('stylesheet', ccsfile_out))
 
     body_node = make_child_node(html_node, 'body')
-    _ = create_plot_tables(body_node, config_info['table_desc'], inputbase, outdir, **kwcopy)
 
-    rafts = get_raft_names_dc(run)
+    kwcopy['raft'] = raft
+    kwcopy['slot'] = slot
+    kwcopy['dataset'] = dataset
+    kwcopy['dataid'] = dict(raft=raft, slot=slot, dataset=dataset)
+    create_plot_tables(body_node, config_info['table_desc'], inputbase, outdir, **kwcopy)
 
-    for raft in rafts:
-        dataid = dict(run=run, raft=raft)
-        for slot in ALL_SLOTS:
-            dataid['slot'] = slot
-            write_slot_report(dataid, inputbase, outbase, **kwargs)
-        write_raft_report(dataid, inputbase, outbase, **kwargs)
+    write_tree_to_html(html_node, html_file)
+
+
+def write_summary_report_by_raft(dataset, raft, inputbase, outbase, **kwargs):
+    """Create table with descriptions and a plots
+
+    Parameters
+    ----------
+    dataset : `str`
+        Dataset id
+    raft : `str`
+        Raft id
+    inputbase : `str`
+        Input base directory
+    outbase : `str` or `None`
+        Output directory
+
+    """
+    kwcopy = kwargs.copy()
+
+    config_info = get_report_config_info('summary_by_raft_plot_tables', **kwcopy)
+
+    kwcopy.update(config_info['defaults'])
+
+    if outbase is None:
+        outdir = None
+        html_file = None
+    else:
+        outdir = os.path.join(outbase, raft)
+        html_file = os.path.join(outdir, '%s_index.html' % dataset)
+        makedir_safe(html_file)
+        ccsfile_out = handle_file(config_info['cssfile'], outdir, action='copy')
 
     kwcopy['html_file'] = html_file
-    kwcopy['rafts'] = rafts
-    if rafts:
-        create_raft_table(body_node, **kwcopy)
-        write_tree_to_html(html_node, html_file)
-    else:
-        sys.stdout.write("No data, skipping run\n")
+    sys.stdout.write("Writing report for %s:%s to %s\n" % (dataset, raft, html_file))
+
+    html_node = ET.Element('html')
+    create_report_header(html_node,
+                         title="%s %s results" % (dataset, raft),
+                         stylesheet=kwcopy.pop('stylesheet', ccsfile_out))
+
+    body_node = make_child_node(html_node, 'body')
+
+    kwcopy['raft'] = raft
+    kwcopy['dataset'] = dataset
+    kwcopy['dataid'] = dict(raft=raft, dataset=dataset)
+    create_plot_tables(body_node, config_info['table_desc'], inputbase, outdir, **kwcopy)
+
+    for slot in ALL_SLOTS:
+        write_summary_report_by_slot(dataset, raft, slot, inputbase, outbase, **kwargs)
+
+    create_slot_table(body_node, prefix="%s_" % dataset, **kwcopy)
+
+    write_tree_to_html(html_node, html_file)
 
 
 
@@ -715,12 +879,10 @@ def write_summary_report(dataset, inputbase, outbase, **kwargs):
     """
     kwcopy = kwargs.copy()
 
-    sys.stdout.write("Writing report for %s\n" % dataset)
-
     config_info = get_report_config_info('summary_plot_tables', **kwcopy)
 
     kwcopy.update(config_info['defaults'])
-    kwcopy['dataid'] = dict(dataset=dataset)
+    rafts = kwcopy.get('rafts', NINE_RAFTS)
 
     if outbase is None:
         outdir = None
@@ -731,14 +893,28 @@ def write_summary_report(dataset, inputbase, outbase, **kwargs):
         makedir_safe(html_file)
         ccsfile_out = handle_file(config_info['cssfile'], outdir, action='copy')
 
+    kwcopy['html_file'] = html_file
+    sys.stdout.write("Writing report for %s to %s\n" % (dataset, html_file))
+
     html_node = ET.Element('html')
     create_report_header(html_node,
-                         title="%s results" % dataset,
+                         title="Summary results for dataset %s" % dataset,
                          stylesheet=kwcopy.pop('stylesheet', ccsfile_out))
 
     body_node = make_child_node(html_node, 'body')
 
+    kwcopy['dataid'] = dict(dataset=dataset)
     create_plot_tables(body_node, config_info['table_desc'], inputbase, outdir, **kwcopy)
+
+    for raft in rafts:
+        kwcopy.pop('dataid', None)
+        write_summary_report_by_raft(dataset, raft, inputbase, outbase,
+                                     h3_text="Summary results by Raft",
+                                     **kwcopy)
+
+    kwcopy['rafts'] = rafts
+    kwcopy['dataid'] = dict(dataset=dataset)
+    create_raft_table(body_node, prefix="%s_" % dataset, **kwcopy)
 
     create_run_table(body_node, dataset, **kwcopy)
